+++ packages/testing/playwright/pages/CanvasPage.ts
@@ -1,20 +1,15 @@
-import type { Locator, Page } from '@playwright/test';
+import type { Locator } from '@playwright/test';
 import { nanoid } from 'nanoid';
 
 import { BasePage } from './BasePage';
 import { ROUTES } from '../config/constants';
 import { resolveFromRoot } from '../utils/path-helper';
+import { LogsPanel } from './components/LogsPanel';
 import { StickyComponent } from './components/StickyComponent';
-import { LogsPanel } from './parts/LogsPanel';
 
 export class CanvasPage extends BasePage {
-	readonly logsPanel: LogsPanel;
 	readonly sticky = new StickyComponent(this.page);
-
-	constructor(page: Page) {
-		super(page);
-		this.logsPanel = new LogsPanel(this.page.getByTestId('logs-panel'));
-	}
+	readonly logsPanel = new LogsPanel(this.page.getByTestId('logs-panel'));
 
 	saveWorkflowButton(): Locator {
 		return this.page.getByRole('button', { name: 'Save' });
